Moved & renamed files, bit of cleanup
