+++ DAQ_System/Source/Platform/component_interface_bridge.hpp
@@ -10,14 +10,14 @@
 #ifndef COMPONENT_INTERFACE_BRIDGE_H
 #define COMPONENT_INTERFACE_BRIDGE_H
 
-#include "Adapter/Interfaces/ilinear_potentiometer.hpp"
-#include "Adapter/LinearPotentiometer/linear_potentiometer_sls1322.hpp"
-
-#include "Application/I_Data_Logger.hpp"
-#include "Application/DataLogger/SdDataLogger.hpp"
-
-#include "daq_pin_map.hpp"
+// C/C++ Standard Libraries
 #include <memory>
+
+// DFR Custom Dependancies
+#include "Adapter/LinearPotentiometer/linear_potentiometer_sls1322.hpp"
+#include "Application/DataLogger/sd_data_logger.hpp"
+#include "daq_pin_map.hpp"
+
 
 
 namespace platform {
